Added the possibility to show the delta to the step goal instead of current steps
Instead of showing the steps you walked it is now also possible to show the steps you need to walk to reach your daily step goal. As long as you are below the step goal the color of the steps will be dark red and as soon as you reach the step goal it will become black.